+++ test/runtests.jl
@@ -1,2 +1,1 @@
-module util_tests include("util.jl") end
 module ais_tests include("ais.jl") end